alerts-consumer/requirements.txt
kafka-python
requests
psycopg2-binary
numpy
scikit-learn